Add a flag to Calloc to allow the caller to specify the usage of the byte slice.
